requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
fastapi==0.115.4
firecrawl-py==1.4.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.0
loguru==0.7.2
nest-asyncio==1.6.0
openai==1.53.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.41.2
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==13.1
annotated-types==0.7.0
anyio==4.6.2.post1
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
fastapi==0.115.4
firecrawl-py==1.4.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.0
langfuse==2.53.9
loguru==0.7.2
nest-asyncio==1.6.0
openai==1.53.1
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.41.2
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==13.1
wrapt==1.16.0
